requirements.txt
aioredis==1.3.1
asgiref==3.2.3
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.2.1
Automat==20.2.0
certifi==2019.11.28
cffi==1.14.0
channels==2.3.0
channels-redis==2.4.2
chardet==3.0.4
constantly==15.1.0
cryptography==2.8
daphne==2.4.1
Django==2.0.2
django-cors-headers==3.2.1
django-tenant-schemas==1.10.0
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
djangorestframework-simplejwt==4.4.0
fuzzywuzzy==0.18.0
hiredis==1.0.1
html5lib==1.0.1
hyperlink==19.0.0
idna==2.9
incremental==17.5.0
msgpack==0.6.2
ordered-set==3.1.1
Pillow==7.0.0
psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.19
PyHamcrest==2.0.0
PyJWT==1.7.1
pyOpenSSL==19.1.0
PyPDF2==1.26.0
pytz==2019.3
reportlab==3.5.34
requests==2.23.0
service-identity==18.1.0
six==1.14.0
sqlparse==0.3.0
stripe==2.42.0
Twisted==19.10.0
txaio==20.1.1
urllib3==1.25.8
webencodings==0.5.1
xhtml2pdf==0.2.4
xlrd==1.2.0
zope.interface==4.7.1
